Addressed comments -- minor tweaks and file refactoring
