+++ metaflow/datastore/s3_backend.py
@@ -1,8 +1,9 @@
 import os
 
 from ..metaflow_config import DATASTORE_SYSROOT_S3
-from .common import DataException
 from .datastore_backend import DataStoreBackend
+from .exceptions import DataException
+
 
 try:
     # python2
